KerrKS is not WIP anymore
Modify the WIP (work in progress) class constructor to not issue a
warning if passed the empty string instead of a class name. Use this
facility to stop issuing the warning for KerrKS, which has grown in
maturity. This is an ABI-safe change. Next step is to stop inheriting
from WIP, but this will break the ABI and is not for the stable
branch.

The warning when using the specific integrator is now severe. This is
the part that really is still buggy in KerrKS.