Simplify setting arbitrary parameters from the command line
Object::setParameter(name, content, unit) is made recursive: if name
contains "::" (e.g. Astrobj::Spectrum::Temperature), the component
before the first occurence of "::" (here, "Astrobj") is interpreted as
a path. A property by that name in the curent object is sought. If it
is found, setParameter is called on this child object with the
remainder of the string (here "Spectrum::Temperature") and the other
parameters unchanged (content and unit).

On the command-line, replace the bunch of --TYPE-parameter options
with a single --parameter option, which applies to the root (Scenery)
level and uses this facility to access members in the Astrobj, Screen
etc.