Allow versionned name for the Python plug-in
E.g. if $PYTHON is python3.4, then the plug-in will also be called
python3.4 (as in libgyoto-python3.4.so). This way we can build and
install several versions of the plug-in, that will run in the various
interpreters. Note that the plug will not run in python2.7 if compiled
with python3.4.