Introduce python/gyoto_swig.h and gyoto_std_swig.h
This simplifies the maintenance of Python extensions for external
plug-ins: instead of hardcoding the #includes and function prototpes
needed to import gyoto.i and gyoto_std.i, one can just #include the
corresponding header. At some point it would be great to find how to
preprocess the information directly from the .i files, if that is
possible.