Use only major library version in GYOTO_PLUGDIR
Up to know, the versioned directory for plug-ins used the full libtool
versinfo, changing with each new library soname revision.  From know
onwards, it is only the major number, which is kept for
backward-compatible changes. Users will not need to recompile their
plug-ins when upgrading to a backwarsd-compatible version of Gyoto.