Add new astrobj XillverReflection
This astrobj contains two fits table.
The first one gives the illumination over a thin
disk as a function of r and phi, that should be
computed externally for a lamp in Keplerian rotation
at some radius rlamp.
The second one is a direct translation of the
reflection spectra computed with Javier Garcia's
Xillver code, see https://hea-www.cfa.harvard.edu/~javier/xillver/.
The output is the ray-traced reflection spectrum,
with the self-consistent rotating lamp illumination.