Replace CutOffIneV by CutOff (in Hz) in Spectrum::PowerLaw
Rationale: all Gyoto APIs should be in SI. Besides, Band in
Spectrometer is in Hz by default. Cut-off frequencies should be in
the same unit.

The new interface supports unit, so it is still possible to use eV
but also any unit convertible to Hz, m or eV. The user may replace
  <CutOffIneV> 1 2 </CutOffineV>
with
  <CutOff unit='eV'> 1 2 </CutOff>

In addition, document the properties.