Python: nicer syntax for gmunu and christoffel, test Christoffels
A more Pythonic syntax was already in place for Metric::Generic::gmunu
(and others). Here we make sure this nicer syntax remains available
in metrics that reimplement these methods.

Add two functions (christoffel_numerical and check_christoffel) in
gyoto.metric to ease testing of Christoffel symbols.

Use check_christoffel to test all the metrics in stdplug and
RotStar3_1 in lorene.