Metric: ensure consistency between gmunu versions and gmunu_up
Add tests for all metrics in stdplug to check that both versions of
gmunu are equivalent and the inverse of gmunu_up. Make sure all
these methods are callable from Python in a Python and consistent
fashion. Add a flavour of gmunu to ChernSimons for which the KerrBL
implementation was previously exposed.